requirements.txt
# MedGPT - Medical Visual Question Answering System
# Core ML
torch>=2.4.0
transformers>=4.51.0
peft>=0.14.0
accelerate>=1.2.0
bitsandbytes>=0.45.0
trl>=0.15.0

# Qwen3-VL specific
qwen-vl-utils>=0.0.8

# Data
datasets>=3.0.0
Pillow>=10.0.0

# Evaluation
nltk>=3.9.0
rouge-score>=0.1.2

# Logging
wandb>=0.19.0
tensorboard>=2.18.0

# Web Application
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18

# Explainability & Visualization
matplotlib>=3.9.0
opencv-python>=4.10.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.67.0
numpy>=1.26.0
scipy>=1.14.0
